% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{accident_2013}
\alias{accident_2013}
\title{Print "accident_2013"}
\description{
This is a sample of the source data. The source data can be found inside the "data" subfolder and can be read in R with the fars_read function
}

