% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Print "fars_map_state"}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number. Only one value can be passed as state.num (either numeric or character). If not a proper state.num is specified, an error message will be thrown.}

\item{year}{The value of the year. Only one value can be passed as year (either numeric or character).}
}
\value{
This function returns a plot of the incidences for the specified year and state. Specifically, it will return a map of the state specified including the incidences as dots.
}
\description{
This is function that takes as input the state and the year and provides as output a map of the specified state demonstarting the exact location of fatal injuries suffered in motor vehicle traffic crashes (as dots).
}
\note{
Needed packages for this function are: dplyr, maps, graphics. Furthermore, if there are not incidences for the specified combination of year/state number, a relevant message will pop-up.
}
\examples{
fars_map_state(45,2013)

}
