% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Print "fars_read_years"}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The value of the year. The year can be set as character or numeric and will not break the execution of the code. Also you can use a vector of years.}
}
\value{
This function returns 2 columns Month and Year.
}
\description{
This function creates 2 columns Month and Year related to the source file of the specified year.
}
\note{
In this function, are used the mutate ad select functions from dplyr package. So please make sure to have upfront install and load the dplyr package otherwise you may face issues/errors during the execution.
}
\examples{
fars_read_years("2013")
fars_read_years(c(2013,2015))

}
