% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Print "fars_summarize_years"}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector including the years that will be included in the analysis. As above, it can be one year or more and can be passed as character or numeric.}
}
\value{
This function returns a table of tibble class, with one row per month and with at least two columns. One presenting the MONTH and one for each year used as parameter during the execution.It basically provides the number of fatal injuries suffered in motor vehicle traffic crashes by Month and year.
}
\description{
This is a function that, summarizes the number of rows included per month in the source file and the specified year. The format of the results, is a table of tibble class, with one row per month and with at least two columns. One presenting the MONTH and one for each year used as parameter during the execution.
}
\note{
Again for the proper execution of this function, we need to verify that the dplyr package is installed and loaded and also that the source files are located in the folder that is currently set as working directory.
}
\examples{
fars_summarize_years(c("2013","2014"))

}
