% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print "make_filename"}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The value of the year.}
}
\value{
This function returns the name "accident_%d.csv.bz2" but instead of %d will concatenate the value of year specified as input in the function.
}
\description{
This is a function that, returns the name of the file related to the year specified as parameter.
}
\examples{
make_filename(2015) 
will reaturn as output "accident_2015.csv.bz2"

}
